+++ src/common/Form/FileInput/UppyConfig.ts
