+++ src/pages/Events/Content/EventsListings/EventsListings.tsx
@@ -3,7 +3,7 @@ import type { IEvent, IEventDB } from 'src/models/events.models'
 import { Button, MoreContainer } from 'oa-components'
 import { Flex, Box, Heading } from 'theme-ui'
 import EventCard from '../../EventCard/EventCard'
-import TagsSelect from 'src/components/Tags/TagsSelect'
+import TagsSelect from 'src/common/Tags/TagsSelect'
 import { inject, observer } from 'mobx-react'
 import type { EventStore } from 'src/stores/Events/events.store'
 import type { UserStore } from 'src/stores/User/user.store'
