+++ src/pages/Howto/Content/HowtoList/HowtoList.tsx
@@ -3,7 +3,7 @@ import * as React from 'react'
 import { AuthWrapper } from 'src/common/AuthWrapper'
 import { Button, MoreContainer, Loader } from 'oa-components'
 import { Heading, Input, Flex, Box } from 'theme-ui'
-import TagsSelect from 'src/components/Tags/TagsSelect'
+import TagsSelect from 'src/common/Tags/TagsSelect'
 import { VirtualizedFlex } from 'src/pages/Howto/VirtualizedFlex/VirtualizedFlex'
 import type { HowtoStore } from 'src/stores/Howto/howto.store'
 import type { UserStore } from 'src/stores/User/user.store'
